move layers.py
